requirements.txt
Django>=2.0
psycopg2
celery==4.2.1
factory-boy==2.11.1
